Test task04 for lesson01

+++ src/main/java/ru/yandex/practicum/theme22/lesson8internal/TestPracticum2208.java
@@ -6,8 +6,8 @@ import java.util.List;
 public class TestPracticum2208 {
     public static void main(String[] args) {
         List<Double> list1 = new ArrayList<>();
-        List<Float> list2 = new ArrayList<>();
-        Merger<Double, Float> merger = new Merger<>();
+        List<Double> list2 = new ArrayList<>();
+        Merger<Double, Double> merger = new Merger<>();
         merger.merge(list1, list2, 1);
 
     }
